Refactor:Fix:207:reflect testFile after test #207
