+++ front-end/src/router.jsx
@@ -11,7 +11,7 @@ import UserlistPage from "pages/UserlistPage";
 import AuthLayout from "layout/AuthLayout";
 
 /*eslint-disable */
-import FoodFormTest from "pages/FoodFormTest";
+// import FoodFormTest from "pages/FoodFormTest";
 const routerData = [
   {
     id: 0,
@@ -38,16 +38,16 @@ const routerData = [
     id: 3,
     path: "/foodForm",
     label: "음식 입력 페이지",
-    // element: <FoodFormPage />, // 여기도
-    element: <FoodFormTest />, // 여기도
+    element: <FoodFormPage />, // 여기도
+    // element: <FoodFormTest />, // 여기도
     withAuth: false // !!!! 여기 제거 해야 함
   },
   {
     id: 4,
     path: "/foodForm/info/:foodId",
     label: "음식 입력 결과 페이지",
-    // element: <FoodFormPage />,
-    element: <FoodFormTest />,
+    element: <FoodFormPage />,
+    // element: <FoodFormTest />,
     withAuth: false // !!!! 여기 제거 해야 함
   },
   {
